# Improve performance in RomanNumeralExtensions.FromRoman