Here is a checklist you should tick through before submitting a pull request: 
 - [x] Implementation is clean
 - [x] Code adheres to the existing coding standards; e.g. no curlies for one-line blocks, no redundant empty lines between methods or code blocks, spaces rather than tabs, etc.
 - [x] No Code Analysis warnings
 - [x] There is proper unit test coverage
 - [x] If the code is copied from StackOverflow (or a blog or OSS) full disclosure is included. That includes required license files and/or file headers explaining where the code came from with proper attribution
 - [x] There are very few or no comments (because comments shouldn't be needed if you write clean code)
 - [x] Xml documentation is added/updated for the addition/change
 - [x] Your PR is (re)based on top of the latest commits from the `main` branch (more info below)
 - [x] Readme is updated if you change an existing feature or add a new one
 - [x] Run either `build.cmd` or `build.ps1` and ensure there are no test failures


Performance improvement:

| Method |     Mean |      Min |      Max |        Op/s |   Gen0 | Allocated |
|------- |---------:|---------:|---------:|------------:|-------:|----------:|
| Before | 457.1 ns | 452.0 ns | 462.8 ns | 2,187,737.5 | 0.1259 |     264 B |
|  After | 423.9 ns | 423.6 ns | 424.3 ns | 2,359,098.2 | 0.1259 |     264 B |